Prompt for SSH passphrase when necessary
Automatically prompt the user for its passphrase when connecting to a
remote host using an encrypted SSH key and no agent.
The passphrase is remembered until the end of the session.
This mimicks the behavior of fabric v1.

Note that commands called from the fabfile (git, rsync, scp) will still
ask for the passphrase everytime.